feat: old logic on main removed and refactorized to modules and .toml also
