+++ src/input_device_monitor/mod.rs
@@ -1,0 +1,4 @@
+pub mod sender;
+pub mod event;
+pub mod app;
+
