+++ src/input_device_monitor/sender/mod.rs
@@ -1,0 +1,3 @@
+pub mod sender;
+
+pub use sender::{ISender, SocketClientSender};
